rename com.orange namespace to com.orange.servicebroker.staticcreds
