+++ src/main/java/com/orange/Application.java
@@ -10,7 +10,7 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
 import static org.springframework.cloud.servicebroker.model.BrokerApiVersion.API_VERSION_ANY;
 
 @EnableGlobalMethodSecurity
-@SpringBootApplication(scanBasePackages = { "com.orange" })
+@SpringBootApplication(scanBasePackages = { "com.orange.servicebroker.staticcreds" })
 @RefreshScope
 public class Application {
 	public static void main(String[] args) {
