--- src/main/java/com/orange/service/CredsServiceInstanceBindingService.java
@@ -1,34 +1,0 @@
-package com.orange.service;
-
-import com.orange.model.Plan;
-import com.orange.model.PlanRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.servicebroker.model.CreateServiceInstanceBindingRequest;
-import org.springframework.cloud.servicebroker.model.CreateServiceInstanceBindingResponse;
-import org.springframework.cloud.servicebroker.model.DeleteServiceInstanceBindingRequest;
-import org.springframework.cloud.servicebroker.service.ServiceInstanceBindingService;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Service
-public class CredsServiceInstanceBindingService implements ServiceInstanceBindingService {
-    private PlanRepository planRepository;
-
-    @Autowired
-    public CredsServiceInstanceBindingService(PlanRepository planRepository) {
-        this.planRepository = planRepository;
-    }
-
-    @Override
-    public CreateServiceInstanceBindingResponse createServiceInstanceBinding(CreateServiceInstanceBindingRequest request) {
-        UUID planId = UUID.fromString(request.getPlanId());
-        final Optional<Plan> plan = planRepository.find(planId);
-        return new CreateServiceInstanceBindingResponse(plan.map(Plan::getFullCredentials).orElse(null));
-    }
-
-    @Override
-    public void deleteServiceInstanceBinding(DeleteServiceInstanceBindingRequest arg0) {
-    }
-}
