+++ src/main/java/com/orange/servicebroker/staticcreds/domain/PlanRepository.java
@@ -1,0 +1,13 @@
+package com.orange.servicebroker.staticcreds.domain;
+
+import java.util.Optional;
+import java.util.UUID;
+
+/**
+ * Created by YSBU7453 on 04/04/2016.
+ */
+public interface PlanRepository {
+
+    Optional<Plan> find(UUID servicePlanId);
+
+}
