+++ src/main/java/com/orange/servicebroker/staticcreds/domain/ServiceRepository.java
@@ -1,0 +1,12 @@
+package com.orange.servicebroker.staticcreds.domain;
+
+import java.util.List;
+
+/**
+ * Created by YSBU7453 on 04/04/2016.
+ */
+public interface ServiceRepository {
+
+    List<Service> findAll();
+
+}
